+++ packages/sampler/stories/qa/Header.js
@@ -22,7 +22,7 @@ const inputData = {
 	longTitle: 'Core, The building blocks of an Enact application. Moonstone, our TV-centric UI library.',
 	longTitleBelow: 'An app development framework built atop React that’s easy to use, performant and customizable. The goal of Enact is to provide the building blocks for creating robust and maintainable applications.',
 	longSubTitleBelow: 'With over 50 components to choose from, Moonstone provides a solid base for creating applications designed for large screens. The Enact team welcomes contributions from anyone motivated to help out.',
-	longRtlTitle: 'غينيا واستمر العصبة ضرب قد. وباءت الأمريكي الأوربيين هو به،, هو العالم، الثقيلة بال. مع وايرلندا الأوروبيّون كان, قد بحق أسابيع العظمى واعتلاء. انه كل وإقامة المواد.'
+	longRtlTitle: 'هذا النص طويل ولكن ليس طويلاً. بالتأكيد ليست قصيرة جدا ، على الرغم من.'
 };
 
 const headerComponents = <Button>Header Button</Button>;
@@ -177,6 +177,23 @@ storiesOf('Header', module)
 					title={text('title', Config, inputData. longRtlTitle)}
 					titleBelow={text('titleBelow', Config, inputData.shortRtlTitleBelow)}
 					subTitleBelow={text('subTitleBelow', Config, inputData.shortRtlTitleBelow)}
+					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
+				>
+					{addHeaderComponents ? headerComponents : null}
+				</Header>
+			);
+		}
+	)
+	.add(
+		'RTL text, long title, Compact',
+		context => {
+			context.noHeader = true;
+			const addHeaderComponents = boolean('add headerComponents', Config);
+			return (
+				<Header
+					type="compact"
+					title={text('title', Config, inputData. longRtlTitle)}
+					titleBelow={text('titleBelow', Config, inputData.shortRtlTitleBelow)}
 					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
 				>
 					{addHeaderComponents ? headerComponents : null}