+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [3.2.4] - 2019-11-07
+
+### Fixed
+
+- `moonstone/Notification` line height for non-latin locales
+- `moonstone/Notification` to show all buttons in one line
+- `ui/Marquee` text aligment when content is centered
+
 ## [3.2.3] - 2019-11-01
 
 ### Changed