+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [3.2.4] - 2019-11-07
+
+### Fixed
+
+- `moonstone/Notification` line height for non-latin locales
+- `moonstone/Notification` to show all buttons in one line
+
 ## [3.2.3] - 2019-11-01
 
 ### Changed