+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [3.2.4] - 2019-11-07
+
+### Fixed
+
+- `ui/Marquee` text aligment when content is centered
+
 ## [3.2.3] - 2019-11-01
 
 ### Fixed